Single FindParser function for all parsers
